// src/app/atendimento/vestigios/enums/vestigio-categoria.enum.ts
export enum VestigioCategoria {
  Biologicos = 'biologicos',
  Balisticos = 'balisticos',
  Papiloscopicos = 'papiloscopicos',
  ImpressoesMarcas = 'impressoes-marcas',
  Fisicos = 'fisicos',
  Quimicos = 'quimicos',
  Microvestigios = 'microvestigios',
  ComportamentaisContextuais = 'comportamentais-contextuais',
}